Pull out traceability env's from gisday into ngx-environment
Also exports the variables needed to be substituted at image build time to reduce application-specific implementations
