+++ apps/geoservices-angular/src/environments/environment.dev.ts
@@ -2,15 +2,14 @@ export const environment = {
   production: true
 };
 
+export { metadata } from '@tamu-gisc/common/ngx/environment';
+
 export const api_url = `https://nodes.geoservices.tamu.edu/api/gsvcs-transport/`;
 export const legacy_host = 'https://geoservices-dev.geoservices.tamu.edu';
 export const legacy_api_url = `https://geoservices-dev.geoservices.tamu.edu/rest/`;
 export const accounts_url = `https://geoservices-dev.geoservices.tamu.edu`;
 export const geoprocessing_api_host_override = `https://geoservices-dev.geoservices.tamu.edu/api`;
 
-export const release_id = '___RELEASE_ID___';
-export const machine_name = '___MACHINE_NAME___';
-export const environment_mode = '___ENVIRONMENT_MODE___';
 export const turnstile_sitekey = '___TURNSTILE_SITE_KEY___';
 
 export * from './definitions';
