+++ apps/geoservices-angular/src/environments/environment.prod.ts
@@ -2,15 +2,14 @@ export const environment = {
   production: true
 };
 
+export { metadata } from '@tamu-gisc/common/ngx/environment';
+
 export const api_url = `___API_URL___`;
 export const legacy_host = '___LEGACY_HOST___';
 export const legacy_api_url = `___LEGACY_API_URL___`;
 export const accounts_url = `___ACCOUNTS_URL___`;
 export const geoprocessing_api_host_override = `___GEOPROCESSING_API_HOST_OVERRIDE___`;
 
-export const release_id = '___RELEASE_ID___';
-export const machine_name = '___MACHINE_NAME___';
-export const environment_mode = '___ENVIRONMENT_MODE___';
 export const turnstile_sitekey = '___TURNSTILE_SITE_KEY___';
 
 export * from './definitions';
