+++ libs/common/ngx/environment/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/common-ngx-environment.module';
 
 export * from './lib/services/environment.service';
+export * from './lib/constants/build-traceability.constants';
