+++ libs/geoservices/ngx/src/lib/core/components/footer/release-info/release-info.component.html
@@ -1,5 +1,5 @@
 <div class="container12">
   <div class="column12">
-    <p><span>{{ (release_meta.args | async)?.machine_name }}</span> | <span *ngIf="release_meta.isRelease | async">{{(release_meta.args | async)?.release}} |</span> {{ (release_meta.args | async)?.environment }}</p>
+    <p><span>{{ release_meta?.nodeName }}</span> | <span>{{(release_meta)?.gitTag}} |</span> {{ release_meta?.environment }}</p>
   </div>
 </div>
